+++ Tests/Obfy.Console.Tests/WizardDefaultsTests.cs
@@ -65,4 +65,49 @@ public class WizardDefaultsTests
 
         context.Settings.SymbolRenaming.PreserveXaml.ShouldBeTrue();
     }
+
+    [Fact]
+    public void ApplyUseCaseDefaults_ClassLibrary_PreservesPublicApi()
+    {
+        var context = new WizardContext
+        {
+            UseCase = "Class Library / NuGet Package",
+            Settings = new ObfySettings()
+        };
+
+        ConfigurationWizard.ApplyUseCaseDefaults(context);
+
+        context.Settings.SymbolRenaming.PreservePublicApi.ShouldBeTrue();
+    }
+
+    [Fact]
+    public void ApplyUseCaseDefaults_AspNet_ExcludesMvcAttributes()
+    {
+        var context = new WizardContext
+        {
+            UseCase = "Web Application (ASP.NET)",
+            Settings = new ObfySettings()
+        };
+
+        ConfigurationWizard.ApplyUseCaseDefaults(context);
+
+        context.Settings.Exclusions.Attributes.ShouldContain("Microsoft.AspNetCore.Mvc.RouteAttribute");
+        context.Settings.Exclusions.Attributes.ShouldContain("Microsoft.AspNetCore.Mvc.ApiControllerAttribute");
+    }
+
+    [Fact]
+    public void ApplyUseCaseDefaults_IsPublicApi_PreservesPublicNames()
+    {
+        var context = new WizardContext
+        {
+            UseCase = "Desktop Application",
+            IsPublicApi = true,
+            Settings = new ObfySettings()
+        };
+
+        ConfigurationWizard.ApplyUseCaseDefaults(context);
+
+        context.Settings.SymbolRenaming.PreservePublicApi.ShouldBeTrue();
+        context.Settings.SymbolRenaming.PreserveXaml.ShouldBeTrue();
+    }
 }